Animation Work
Did some simple animation work to set the fundamentals for when we get out platforming transition and when we get our sprites for the main character.

NOTE: Need to figure out how to set up a proper way to detect jumping and make it trigger the animator parameters so that i can properly sync up animations.